덧글 기능 구현
1. 덧글 삭제 및 추가 기능 구현
2. 덧글 작성자에 마이페이지 링크 바인딩
3. 덧글 관련 서비스 및 컨트롤러 작성

******* To Do ******
1. 태그 기능 구현 필요
2. 무한 스크롤 기능 구현(로딩 속도가 너무 느림)
3. 최신순으로 덧글 정렬(시간은 넘겨주는데 형식이 조금 틀림)
4. UI 개선
5. 마이페이지 덧글 뿌려주기

******* 발생 이슈 ******
1. 모든 댓글을 처음부터 load하니 너무 많은 쿼리가 날라가 로딩속도에 문제가 있음 ==> 무한 스크롤 또는 덧글을 클릭시만 요청

+++ src/main/java/com/team2/webservice/sprint1/jpa/CommentRepository.java
@@ -3,5 +3,6 @@ package com.team2.webservice.sprint1.jpa;
 import com.team2.webservice.sprint1.vo.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
+
 } // 인터페이스를 생성후, JpaRepository<Entity 클래스, PK타입>을 상속하면 기본적인 CRUD 메소드가 자동생성됨
